requirements.txt
google-genai>=1.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
gspread>=6.0.0
python-dotenv>=1.0.0
pandas>=2.2.0
tenacity>=8.3.0
rich>=13.7.0
streamlit>=1.40.0
